+++ components/utils.ts
@@ -40,3 +40,10 @@ export const getCurrentWeek = (dates: string[]) => {
 
   return currentWeek;
 };
+
+export const getPlayerName = (playerId, players) => {
+  if (!playerId) return;
+  const searchedPlayer = players.find((player) => playerId == player.id);
+
+  return `${searchedPlayer.user.firstName} ${searchedPlayer.user.lastName}`;
+};
